Add support for hosting arbitrary "media" files
Note that we've gone with a simple approach of just copying all files in a particular folder. This will cover future use cases of media linked from blog posts (where we could have also looked on demand for linked files, like we do for Markdown links), but also files which we'd like to host for any other reason, e.g. for sharing or for a profile picture for a forum which doesn't host them itself (CaB).

One obvious future extension of this is to allow for a hierarchical structure, copying the directory recursively.